+++ Emulator/Cartridges/CartridgeRom.cpp
@@ -60,8 +60,8 @@ CartridgeRom::_load(const u8 *buffer)
     // Read packet data
     for (int i = 0; i < size; i++) rom[i] = util::read8(reader.ptr);
 
-    trace(SNP_DEBUG, "Recreated from %ld bytes\n", reader.ptr - buffer);
-    return reader.ptr - buffer;
+    trace(SNP_DEBUG, "Recreated from %ld bytes\n", isize(reader.ptr - buffer));
+    return isize(reader.ptr - buffer);
 }
 
 isize
@@ -74,8 +74,8 @@ CartridgeRom::_save(u8 *buffer)
     // Write packet data
     for (int i = 0; i < size; i++) util::write8(writer.ptr, rom[i]);
 
-    trace(SNP_DEBUG, "Serialized to %ld bytes\n", writer.ptr - buffer);
-    return writer.ptr - buffer;
+    trace(SNP_DEBUG, "Serialized to %ld bytes\n", isize(writer.ptr - buffer));
+    return isize(writer.ptr - buffer);
 }
 
 bool
