+++ Emulator/Media/AnyCollection.h
@@ -29,10 +29,10 @@ public:
     virtual PETName<16> itemName(isize nr) const = 0;
     
     // Returns the size of a certain in bytes
-    virtual u64 itemSize(isize nr) const = 0;
+    virtual isize itemSize(isize nr) const = 0;
         
     // Reads a byte from a certain item
-    virtual u8 readByte(isize nr, u64 pos) const = 0;
+    virtual u8 readByte(isize nr, isize pos) const = 0;
 
     
     //
@@ -40,8 +40,8 @@ public:
     //
 
     // Reads a word from a certain item in little endian or big endian format
-    u16 readWordBE(isize nr, u64 pos) const;
-    u16 readWordLE(isize nr, u64 pos) const;
+    u16 readWordBE(isize nr, isize pos) const;
+    u16 readWordLE(isize nr, isize pos) const;
     
     // Return the load address of this item
     u16 itemLoadAddr(isize nr) const;
