+++ Emulator/Media/Folder.h
@@ -64,8 +64,8 @@ public:
     PETName<16> collectionName() override;
     isize collectionCount() const override;
     PETName<16> itemName(isize nr) const override;
-    u64 itemSize(isize nr) const override;
-    u8 readByte(isize nr, u64 pos) const override;
+    isize itemSize(isize nr) const override;
+    u8 readByte(isize nr, isize pos) const override;
     void copyItem(isize nr, u8 *buf, u64 len, u64 offset) const override;
     
     //
