+++ Emulator/Media/PRGFile.h
@@ -62,6 +62,6 @@ private:
     PETName<16> collectionName() override;
     isize collectionCount() const override;
     PETName<16> itemName(isize nr) const override;
-    u64 itemSize(isize nr) const override;
-    u8 readByte(isize nr, u64 pos) const override;
+    isize itemSize(isize nr) const override;
+    u8 readByte(isize nr, isize pos) const override;
 };
